switch firebase files to ts
