+++ frontend/next.config.js
@@ -4,6 +4,6 @@ const {i18n} = require('./next-i18next.config');
 const nextConfig = {
   reactStrictMode: true,
   i18n,
-}
+};
 
-module.exports = nextConfig
+module.exports = nextConfig;
