--- frontend/utils/authentication.js
@@ -1,16 +1,0 @@
-import { getAuth, signInWithEmailAndPassword } from "firebase/auth";
-
-const auth = getAuth();
-async function signIn(email, password){
-signInWithEmailAndPassword(auth, email, password)
-  .then((userCredential) => {
-    // Signed in 
-    return userCredential.user;
-    // ...
-  })
-  .catch((error) => {
-      throw error;
-  });
-}
-
-module.exports = {signIn}
